- [ZBX-650] Code review (memory leaks, opened file descriptors ...)    [svn merge svn://svn.zabbix.com/branches/1.6 -r6521:6530]
git-svn-id: svn://svn.zabbix.com/trunk@6531 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ src/libs/zbxdbhigh/host.c
@@ -1143,7 +1143,7 @@ static int	DBdelete_graph(
 		{ /* recursion */
 			ZBX_STR2UINT64(graphid, element_data[0]);
 			if( SUCCEED != (result = DBdelete_graph(graphid)) )
-				return result;
+				break;
 		}
 
 		DBfree_result(db_elements);
@@ -1371,11 +1371,12 @@ static int	DBupdate_graph_with_items(
 		}
 
 		if ( SUCCEED != result )
-		{
-			return result;
-		}
+			break;
 	}
 	DBfree_result(chd_graphs);
+
+	if (SUCCEED != result)
+		return result;
 
 	DBexecute("delete from graphs_items where graphid=" ZBX_FS_UI64, graphid);
 
@@ -3175,15 +3176,15 @@ static char*	DBimplode_exp (
 			/* determine key and function */
 			p = parameter-1; /* position of '(' character */
 			*p = '\0';
-			do { /* simple try for C */
-				key = host + host_len + 1;
-				if( NULL == (p = strrchr(key, '.'))) break; /* simpla throw for C */
-				key_len = p - key;
 
-				function = p + 1;
-				function_len = parameter - 1 - function;
+			key = host + host_len + 1;
+			if( NULL == (p = strrchr(key, '.'))) break; /* simpla throw for C */
+			key_len = p - key;
 
-			} while(0); /* simpla finally for C */
+			function = p + 1;
+			function_len = parameter - 1 - function;
+
+			p = parameter-1; /* position of '(' character */
 			*p = '(';
 
 		} while(0);/* simpla finally for C */
